+++ packages/expo/src/Expo.ts
@@ -24,7 +24,6 @@ import * as AR from './AR';
 import * as Brightness from 'expo-brightness';
 import * as FileSystem from 'expo-file-system';
 import * as Google from './Google/Google';
-import * as Haptic from './Haptic/Haptic';
 import * as IntentLauncherAndroid from './IntentLauncherAndroid/IntentLauncherAndroid';
 import * as ScreenOrientation from './ScreenOrientation/ScreenOrientation';
 import * as StoreReview from './StoreReview/StoreReview';
@@ -57,6 +56,8 @@ export { GLView } from 'expo-gl';
 export { GoogleSignIn } from 'expo-google-sign-in';
 import * as ImageManipulator from 'expo-image-manipulator';
 export { ImageManipulator };
+import * as Haptics from 'expo-haptics';
+export { Haptics };
 import * as ImagePicker from 'expo-image-picker';
 export { ImagePicker };
 export { LocalAuthentication } from 'expo-local-authentication';
@@ -81,7 +82,6 @@ export { default as DangerZone } from './DangerZone';
 export { default as ErrorRecovery } from './ErrorRecovery/ErrorRecovery';
 export { Facebook };
 export { Google };
-export { Haptic };
 export { default as Icon } from './Icon';
 export { IntentLauncherAndroid };
 export { default as KeepAwake, activate, deactivate } from 'expo-keep-awake';
@@ -120,6 +120,13 @@ Object.defineProperties(exports, {
       } else {
         return require('expo-sensors').Pedometer;
       }
+    },
+  },
+  Haptic: {
+    enumerable: false,
+    get() {
+      console.log('Module name `Haptic` is deprecated. Use `Haptics` instead.');
+      return Haptics;
     },
   },
 });